Add files via upload
@CODSOFT #codsoft #MachineLearning #internship
Introducing you my Credit Card Fraud Detection Model.
Problem Statement:

The Credit Card Fraud Detection Problem includes modeling past credit card transactions with the knowledge of the ones that turned out to be a fraud. This model is then used to identify whether a new transaction is fraudulent or not.

This is Task 2 for my CodSoft Machine Learning internship.